backend/requirements.txt
fastapi 
uvicorn 
python-dotenv 
pdfminer.six 
python-docx 
requests
pydantic